Added superuser, furthered work on authorization
The first user that is created is now a superuser. The role creation
endpoint requires the role "create:role". The auth middleware has been
adjusted to add the claims into the event context, the accessToken has
been expanded in order to contain claims pertaining to the users
superUser status and permissions